+++ Cargo.toml
@@ -88,14 +88,14 @@ op-alloy-rpc-types-engine = { version = "0.6.8", default-features = false }
 op-alloy-network = { version = "0.6.8", default-features = false }
 
 # sp1
-sp1-lib = { version = "3.0.0", features = ["verify"] }
-sp1-zkvm = { version = "3.0.0", features = ["verify"] }
+sp1-lib = { version = "3.3.0", features = ["verify"] }
+sp1-zkvm = { git = "https://github.com/succinctlabs/sp1-wip.git", branch = "ratan/allocators", features = ["verify", "embedded", "libm", "lib"], default-features = false }
 # Note: This rev includes the SP1 mock groth16 bytes fix, skipping deferred verification, executor opts and network-v2 changes.
-# ratan/v3.0.0-w-network-v2 branch includes the network-v2 changes from `19ae3e1`.
-sp1-sdk = { git = "https://github.com/succinctlabs/sp1.git", rev = "447fda5f759908bbc3091979bd933ee9d126873b", features = [
+# ratan/v3.3.0-network-v2-with-features branch includes the network-v2 changes from `19ae3e1`.
+sp1-sdk = { git = "https://github.com/succinctlabs/sp1.git", rev = "cd23301f627370eace27a6cb28450d65151cbed4", features = [
     "network-v2",
 ] }
-sp1-build = { version = "3.0.0" }
+sp1-build = { version = "3.3.0" }
 
 [profile.release-client-lto]
 inherits = "release"
@@ -106,6 +106,11 @@ lto = "fat"
 [patch.crates-io]
 tiny-keccak = { git = "https://github.com/sp1-patches/tiny-keccak", tag = "tiny_keccak-v2.0.2-patch-v1" }
 sha2 = { git = "https://github.com/sp1-patches/RustCrypto-hashes", package = "sha2", tag = "sha2-v0.10.8-patch-v1" }
-ecdsa = { git = "https://github.com/sp1-patches/signatures", tag = "ecdsa-v0.16.8-patch-v1" }
+ecdsa = { git = "https://github.com/sp1-patches/signatures", tag = "ecdsa-v0.16.9-patch-v3.3.0" }
 substrate-bn = { git = "https://github.com/sp1-patches/bn", tag = "substrate_bn-v0.6.0-patch-v2" }
 sha3 = { git = "https://github.com/sp1-patches/RustCrypto-hashes", package = "sha3", tag = "sha3-v0.10.8-patch-v1" }
+sp1-lib = { git = "https://github.com/succinctlabs/sp1-wip.git", branch = "ratan/allocators" }
+# Note: Have to rename this package to sp1_bls12_381 because it was published with that crate name to kzg_rs. Are people normally importing this crate directly from crates.io?
+# Or are they importing it from sp1-patches? 
+# If from crates.io, then we can remove the patch when this is published.
+sp1_bls12_381 = { git = "https://github.com/sp1-patches/bls12_381.git", branch = "ratan/rename-package" }
