+++ scripts/utils/bin/cost_estimator.rs
@@ -7,12 +7,7 @@ use op_succinct_host_utils::{
     block_range::{
         get_rolling_block_range, get_validated_block_range, split_range_based_on_safe_heads,
         split_range_basic, SpanBatchRange,
-    },
-    fetcher::{CacheMode, OPSuccinctDataFetcher, RunContext},
-    get_proof_stdin,
-    stats::ExecutionStats,
-    witnessgen::run_native_data_generation,
-    ProgramType,
+    }, fetcher::{CacheMode, OPSuccinctDataFetcher, RunContext}, get_proof_stdin, stats::ExecutionStats, witnessgen::run_native_data_generation, HostExecutorArgs, ProgramType
 };
 use op_succinct_scripts::HostExecutorArgs;
 use rayon::iter::{IndexedParallelIterator, IntoParallelRefIterator, ParallelIterator};
