+++ utils/host/src/lib.rs
@@ -8,6 +8,7 @@ pub mod witnessgen;
 use alloy::sol;
 use alloy_consensus::Header;
 use alloy_primitives::B256;
+use clap::Parser;
 use kona_host::{
     kv::{DiskKeyValueStore, MemoryKeyValueStore},
     HostCli,
@@ -17,7 +18,7 @@ use op_succinct_client_utils::{
     boot::BootInfoStruct, types::AggregationInputs, BootInfoWithBytesConfig, InMemoryOracle,
 };
 use sp1_sdk::{HashableKey, SP1Proof, SP1Stdin};
-use std::{fs::File, io::Read};
+use std::{fs::File, io::Read, path::PathBuf};
 
 use anyhow::Result;
 
@@ -132,4 +133,33 @@ pub fn get_agg_proof_stdin(
     stdin.write_vec(headers_bytes);
 
     Ok(stdin)
+}
+
+/// The arguments for the host executable.
+#[derive(Debug, Clone, Parser)]
+pub struct HostExecutorArgs {
+    /// The start block of the range to execute.
+    #[clap(long)]
+    pub start: Option<u64>,
+    /// The end block of the range to execute.
+    #[clap(long)]
+    pub end: Option<u64>,
+    /// The number of blocks to execute in a single batch.
+    #[clap(long, default_value = "10")]
+    pub batch_size: u64, 
+    /// Use cached witness generation.
+    #[clap(long)]
+    pub use_cache: bool,
+    /// Use a fixed recent range.
+    #[clap(long)]
+    pub rolling: bool,
+    /// The number of blocks to use for the default range.
+    #[clap(long, default_value = "100")]
+    pub default_range: u64,
+    /// The environment file to use.
+    #[clap(long, default_value = ".env")]
+    pub env_file: PathBuf,
+    /// Prove flag.
+    #[clap(long)]
+    pub prove: bool,
 }
